clean up the apply club api code
